% Part of the oro.pet package for R
% Distributed under the BSD license: see oro.pet/COPYING
%
% $Id: $

\name{LeanBodyMass}
\Rdversion{1.1}
\alias{leanBodyMass}
\title{Calculating the Lean Body Mass}
\description{
  ...
}
\usage{
leanBodyMass(height, weight, gender = "male")
}
\arguments{
  \item{height}{in cm.}
  \item{weight}{in kg.}
  \item{gender}{should be \dQuote{male} or \dQuote{female}.}
}
\details{
  ...
}
\value{
  ...
}
% \references{}
\author{Brandon Whitcher \email{bjw34032@users.sourceforge.net}}
\seealso{
  \code{\link{standardUptakeValue}}
}
% \examples{}
% \keyword{ ~kwd1 }

