% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{.petWrapper}
\alias{.petWrapper}
\title{Wrapper for oro.pet functions}
\usage{
.petWrapper(name, ...)
}
\arguments{
\item{name}{name of function (without leading ".")}

\item{...}{Additional arguments passed to \code{oro.nifti::.wrapper}}
}
\description{
Simple wrapper for getting functions from
}
