\name{StandardUptakeValue}
\Rdversion{1.1}
\alias{StandardUptakeValue}
\alias{HotSpotSUV}
\alias{TotalSUV}
\title{Calculating Standard Uptake Values for PET}
\description{
  ...
}
\usage{
StandardUptakeValue(data, mask, dose, mass)
HotSpotSUV(slice, radius)
TotalSUV(suv, mask, z, bg, local=TRUE, mname=NULL, nt=NULL)
}
\arguments{
  \item{data}{}
  \item{mask}{}
  \item{dose}{}
  \item{mass}{}
  \item{slice}{}
  \item{radius}{}
  \item{suv}{}
  \item{z}{}
  \item{bg}{}
  \item{local}{}
  \item{mname}{}
  \item{nt}{}
}
\details{
  ...
}
\value{
  ...
}
% \references{}
\author{Brandon Whitcher \email{bjw34032@users.sourceforge.net}}
% \seealso{}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{}

