\name{LeanBodyMass}
\Rdversion{1.1}
\alias{LeanBodyMass}
\title{Calculating the Lean Body Mass}
\description{
  ...
}
\usage{
LeanBodyMass(height, weight, gender = "male")
}
\arguments{
  \item{height}{}
  \item{weight}{}
  \item{gender}{}
}
\details{
  ...
}
\value{
  ...
}
% \references{}
\author{Brandon Whitcher \email{bjw34032@users.sourceforge.net}}
% \seealso{}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% \keyword{ ~kwd1 }

